server/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
apscheduler==3.10.4
pymongo==4.6.0
validators==0.22.0
tldextract==5.1.1
openai>=1.0.0
